Revert "Improve SSH logging for debugging"
